[programmability] connect authority to adapter
Lots more to be done here, but this is a start at finally connecting these two key components!

- Add `Storage` trait + implement it for `AuthorityState`
- Implement the `ResourceResolver` and `ModuleResolver` traits for `AuthorityState` so it can be used as a DB by the Move VM
- Generalize adapter to operate over `Storage` trait rather than `FastXStateView`
- Add logic for handling `MoveCall` order type in the authority
- Add e2e test that creates a `MoveCall` order invoking a nonexistent module that fails appropriately

Note: this temporarily breaks the CLI version of the fastX adapter. It can be fixed by implementing the `Storage` trait for `FastXStateView` + doing a few other things, but I left it broken for now since this PR is already quite meaty.