[programmability] processing logic for module publishing
Filling out the logic for processing a module publish message. This will allow us to define a fastX genesis transaction containing an initial set of modules, which we need for testing anything with programmability.

This is a bit rough (see the many inline TODO's/tasks), but an important step toward getting a system w/ programmability working end-to-end!

- Added the `TxContext` type, which knows how to generate globally unique ID's from the digest of the current transaction.
- Add `adapter::publish`, which deserializes/verifies a set of input modules, generates fresh object ID's for them, and returns fastX objects containing each module
- Filled out the authority code path that handles a module publish message, did some refactoring to make sure the modules get persisted as output objects
- Added an end-to-end test demonstrating successful publishing of a module.