[RFC] decoupling sequence numbers from replay protection
In FastX, replay protection comes from the presence of all `ObjectRef` inputs in the `LockMap`, not from sequence numbers. This is an experimental PR that takes steps toward removing all sequence number-related checks from the authority code to demonstrate this.

A follow-up to this PR could redefine `ObjectRef` to remove sequence numbers altogether. This would allow clients to stop tracking sequence numbers locally and sending them over the wire.