[chore] bump move version
Take advantage of https://github.com/diem/move/pull/13, which will allow us to build Move packages concurrently without encountering flakiness due to races in the cache. Split a test that caused flakiness before to demonstrate that it works.

This will also allow us to move `ObjectBasics` out of the framework and into a test package, but I don't want to do that here because there are a few in-flight PR's that touch this module.