fix: remove Hash + PartialEq on ConfirmedOrder and dependent structs
It's a quick and easy way to check that certificates can't be used in a way that assumes their signature set is meaningful.

We might indeed need to compare CertifiedOrders in the future, in which case a note at the point where a `static_assertions` will blow up will offer a guidansce on what to do.

The reason this doesn't already implement the signature-agnostic way of comparing Certificates is to avoid misuse of that `Eq` implementation in a context where we do not have distinct ValidCertifiedOrder and UncheckedCertifiedOrder types.
Imagine you have a list of CertifiedOrders L and an Eq function that's signature agnostic. You also have a deduplicate function on vecs:
- if all certificates are valid, should you call dedup? Probably yes,
- if all certificates have yet to be checked, should you call dedup? Probably not.

Contributes to #266