refactor: Switch logging from log + env_logger to tracing + tracing_subscriber
This switches the code base to [tokio/tracing](https://github.com/tokio-rs/tracing).

1. tracing's notion of Spans allows us to instrument logs with context that lets us make sense of the application's state, even in a concurrent of distributed context. The [announcement on the tokio blog](https://tokio.rs/blog/2019-08-tracing) is a pretty compact intro to the gist of it.
2. This gives us access to the tracing ecosystem, see https://github.com/MystenLabs/narwhal/pull/26 for more details

This PR transforms the "rails" of logging from using log + event_logger to using tracing, and thereby opens the ability to open Spans or use the `#[instrument]` macro.

It keeps full backward-compatibility with the prior logs. It may mess up the fabric benchmarking files, which I'm no longer sure we maintain. If it does, it would
be because of different time printing formats, in particular it's a bit difficult to make tracing produce the wonky timestamps of log.

As a consequence the python scripts may need massaging to handle RFC 3339 instead of ISO8601.