[demos] add Geniteam.move module with desired monster, farm, cosmetic logic
This is my best attempt to understand what Geniteam is doing and make it accessible from Move code.

Helping out Geniteam by writing the Move code for their game. For now, we add the module in the Sui framework because there is not (I think) an easy way to configure genesis to include modules other than the framework ones. Obviously, this module should be moved outside the framework once we can figure that one out;