fix: Remove head-of-line blocking at the BatchManager
In the current design the AuthorityStore picks atomically a consecutive number for each transaction, and then non-atomically sends it over a mpsc channel to the Batch manager along with the digest. Since multiple threads may operate on the AuthorityState in parallel on different tokio tasks, the sequence numbers may not be in order. To fix this the Batch manager tries to wait until the complete in order sequence is received to sequence and batch.
However, what is an AuthorityState crashes? Then the sequence number will not be sent to the Batch Manager. That may produce a gap in the sequence, resulting in the Batch manager buffering for ever waiting for the missing sequence which will never arrive. This is bad.

This PR gives up on sending transactions in order, and sends them in the order provided by the authority store, which is causal.