ci: various cleanups
This patch does some various cleanups to our rust ci pipeline. In
particular it runs linting (clippy, rustfmt) as well as tests using the
compiler toolchain specified inside of the `rust-toolchain` file instead
of overriding it with either `stable` or `nightly` either of which could
cause CI to break when a new toolchain is released to the stable or
nightly channels.

It also moves canarying of a new compiler toolchain from the workflow
that runs on every PR to a 'nightly' Github Action workflow. It also
prefers using the 'beta' channel instead of 'nightly' channel in order
to reduce noise.