config: better allocation of available ports
Its not ideal to attempt to allocate ports for a large number of nodes
up front. Generally the better thing to do would be to always bind to
port 0 and allow from the OS to properly allocate an available port and
then communicate the exact port that a network interface is bound to to
other entities that would like to speak with it inside of a test
environment.

Given the above still requires a bit of work and plumbing to be done the
next best thing is to be a bit more intelligent about how we're
allocating ports when constructing test configs. This patch changes how
ports are allocated by requesting a fresh, available port from the OS by
binding on port 0, immediately establishing a connection, and then
dropping the connection and the listener. This forces the OS to put the
port into the TIME_WAIT state ensuring that the OS won't hand out the
port for some grace period. Then the port can later be used by by
binding on the port using SO_REUSEADDR (which most rust networking
libraries do by default).