config: rework validator configs (#1939)
This patch does a large rework of our validator configs. Specifically it
creates a single `ValidatorConfig` which represents all of the
configuration a Validator needs to start up. It also creates a
light-weight `ValidatorInfo` type which contains information we expect
to know about other validators (stake, public key, network address) and
will eventually be read from on-chain. This is a replacement for the
`AuthorityInfo` type which was a type that all validators have for all
other validators but included some odd information like the on-disk path
to a validator's db which other validators don't need to know about.

This patch also does an initial rework of creating a `NetworkConfig`
which is used when you want to create N validators to run locally.
There's still some additional work that will need to be done to make
this nicer for testing purposes but this is a good start.