ci: remove cargo-udeps job
Recently we've been seeing the cargo-udeps job take upwards of 30+
minutes to run. This has overwhelmingly become the long-poll in CI by
being 3x longer than the next slowest job. The signal we get from this
job vs the time we spend running it has become a non-ideal trade-off and
as such this patch removes the job.