lock-service: fix bug when closing rocksdb database
Frequently when running tests we'll see test failures with the following
error:

    libc++abi: Pure virtual function called!

This is due to the fact that the thread part of the lock-service runs on
may outlive the main thread leading to a race on closing the rocksdb
database vs cleaning up the C++ runtime, causing this segfault to occur
and preventing the database from being able to cleanly close.

This is fixed by converting these threads to be async tokio tasks which
are cleaned up properly when the runtime shuts down.