config: convert to yaml format for configs
Convert from using JSON to using YAML for our config format. The YAML
format is much easier to work with if you ever need to edit config files
by hand since you do not need to worry about not having trailing commas
as well as the ability to include comments in the configs which will be
invaluable once we begin providing configs to 3rd parties.