[sui tools] prototype of Sui surfer testing tool
Implement a simple [type-inhabitation](https://en.wikipedia.org/wiki/Type_inhabitation)-based bot for Sui for end-to-end testing and (possibly) semi-realistic transaction workload generation. Here's how it works:

- Takes a list of package ID's and a wallet config as an input
- Processes the modules in each package in dependency order (leaves of the dep graph first)
- In a loop, attempt to call each declared `script` function in a module, recording each successful (i.e., non-aborting) call. The loop ends when the surfer stops finding new ways to succeed.
- In order to call each function, the surfer looks at each parameter type, generates a random value of that type (if it's a pure type), or looks for a value of the appropriate type in the wallet (if it's an object type). If it can't find a value of the given type, the attempt to call the function fails.

This is limited in a number of ways, and there are lots of exciting directions to extend this (e.g., using static analysis to inhabit types by calling a function that we know will create a value of that type). But it's still reasonably effective as-is, so wanted to go ahead and put up a PR. I tested it by spinning up a local network (it doesn't work with devnet due to API compability issues) and asking it to "surf" the Sui Framework. It was able to call about 18% of the functions.

```
// start local network
cargo run --bin sui-surfer -- --packages 0x2
// lots of intermediate printing
SurferStats { explored_packages: 1, explored_modules: 21, explored_functions: 34, failed_functions: 188 }
```