[mutex_table] Do not block entire mutex_table bucket when waiting for a single lock
Mutex table maintains individual lock per entry in order to allow independent parallelism for each key.

However, current implementation blocks entire bucket if it needs to wait on a single lock inside this bucket (see added unit test that fails without this PR fixes).

Reason is that we hold a map lock while waiting for acquiring individual lock. This works fine with low level but might bite us when we have higher load as it might significantly reduce parallelism on high contention.

This diff drops bucket lock asap allowing to wait for multiple locks in the same bucket in parallel which seem to be in line with how mutex_table was intended to work.

This PR additionally cleans up some `is_err/unwraps` in code replacing it with `if let` pattern.