[mutex_table] Fix acquire_locks deduplication
It is critical for `acquire_locks` to deduplicate when acquiring multiple locks.

Current implementation does not deduplicate properly - it sorts by bucket key, which can retain order of different objects preventing their deduplication.

For example, consider we want to lock `A` and `B` which falls into same lock bucket, and we pass ``[A, B, A]` to acquire_locks.

 If they fall into same bucket, sorting won't reorder them and we will get same `[A, B, A]` after sorting. Example: https://play.rust-lang.org/?version=stable&mode=debug&edition=2021&gist=f2b9dde8d43040baab279645ee4f5035

 If they are not reordered, they will not be de-duplicated(since Vec::dedup only deduplicate consequent elements). This will lead to a deadlock when trying to acquire lock on the same object we already locked.

 I think originally it made sense to use sorting by the lock bucket key, but since #4362 this is not needed and simple unstable_sort will do.