Remove start_paused from tests
It turns out that narwhal does not work well with `start_paused=true`, this PR attempts to remove it from the tests.

Not all of those tests necessarily use narwhal, but I think if those tests pass it is ok to just remove it from all of them since it is redundant.