[sui framework] add decimals to TreasuryCap for Coin<T>
This is important info for wallets etc. to know about a fungible token. I'm not sure if `TreasuryCap` is the best home for this information (e.g., we may want a big registry for this + token name info etc in the future), but it is the only place we have for now.

This change is inspired by user questions about the minimum denomination of SUI.