Add primary->worker Synchronize RPC (MystenLabs/narwhal#1035)
Moves implementation to a linear async function in the handler.
Gets rid of all parts of the synchronizer that previously existed
only to track pending batch requests to other workers.

NOTE: this changes retries to happen only once instead of
perpetually. Additional retries beyond that point will have to be
driven by the primary.