Consensus: calculate basic integrity hash on consensus messages
This PR introduces rolling integrity hash in the consensus handler calculated as following:

```
IntegrityHash(transaction[n]) = hash(IntegrityHash(transaction[n-1]) || transaction[n].bytes)
```

This hash is then stored along with execution indexes and logged for every 100th certificate.

The purpose of this hash is to quickly find if we have any violation on the narwhal side by examining nodes output.

The motivation to include it is:
(a) Since we are likely to start making significant changes to primary code it would be good to have some integrity check
(b) We have some strange NotFound errors on some nodes, while this was not debugged further it potentially can come from incorrect transaction sequence in narwhal and this hash would be able to (dis)prove this theory.

This should be fairly cheap non-cryptographic hash used for debugging purposes only.

Calculating hash is crash resilient as it basically persist ExecutionIndex -> Hash mapping in the db.