narwhal: fix how primary and workers are notified of reconfig events
Prior to this change we would send outside reconfig events to primaries
via the WorkerToPrimary interface. This was sort of a hack due to not
having an admin interface. Now that we have an admin interface we can
leverage that to notify primaries and then have primaries be responsible
for sending these events to all of its own workers.