Switch to system memory allocator by default (#5614)
This switches memory allocator to the system provided allocator for running Sui nodes. jemalloc can be loaded dynamically instead, so the C++ logic will be able to use jemalloc as well. This has the benefits that C++ allocations can be profiled, and there might be less fragmentation.

Sui node docker image is configured with jemalloc profiling enabled by default.