Add ctx argument to sui::table::new example
`sui::table::new` takes a TxContext as an argument. I got a little confused by the documentation here and had to look up the table definition in the source. The `ctx` is a little out-of-context (pun intended) here but it is a commonly used abbreviation for `TxContext`, so hopefully helps communicate that the `TxContext` argument is required.