[checkpoints] Maintain reverse index consensus_position=>transaction
New `consensus_message_order` table is added that records the (first seen) consensus index for each transaction.

In combination with https://github.com/MystenLabs/sui/pull/5773 we will be able to query all the transactions in the range that are going to be basis for forming checkpoint(still need to add casual dependencies and sort them in a separate process).