[2/x][lamport] Get rid of TransactionEffects::is_object_mutated_here
Spotted while working on lamport timestamps.  This function is only
called in one place (`AuthorityAggregator::get_object_info_execute`)
to check that a validator that object info is being read from isn't
Byzantine, but it is only used in Gateway, and benchmarks.  The former
is going away soon, and the latter does not need to worry about
byzantine validators.

Furthermore, the logic in this function for deleted and wrapped
objects does not seem correct:  It asserts that an object is deleted
in effects if it appears in the deleted effects at `version - 1`,
but the logic in `TemporaryStore::delete_object` increments an
object's version when it is deleted, so it should match.

Opting to remove the code to re-execute certificates instead of trying
to make it work with lamport timestamps as there's a good chance it's
not currently working and therefore needed.