[BUGFIX][deps] Explicitly set patched version of sqlx (#6047)
Sui's dependency on `sqlx` needs to be overridden with a patched
version that uses a more recent version of `libsqlite3-sys`.  The
previous approach used a `patch` configuration in the root
`Cargo.toml`.

This works for builds within this project, but seems to break for
builds that originate from outside the workspace, but depend on one of
the workspace crates.

In this scenario, the patch doesn't apply on the dependency on `sqlx`
in `sui-storage`, so it ends up conflicting with the dependency in
`workspace-hack`.

The most common example of this is someone depending on the Rust
SDK (`sui-sdk`).

This diff ditches the `patch` and sets the dependency override
explicitly on `sui-storage`.

Test Plan:

Sui builds and runs:

```
sui$ cargo build
sui$ cargo simtest
sui$ cargo nextest run
```

Dependents on the Sui SDK build and run, e.g. following the
instructions at https://docs.sui.io/build/rust-sdk succeeds.

Closes #5887