[RFC] [checkpoint v2] Checkpoint notification mechanism
This PR proposes mechanism for in-process consumers to be notified about checkpoints being formed locally.

The first service to likely use this mechanism is the new indexer.

The idea is that consumer gets a stream of `(summary, content)` and when given checkpoint summary is processed, records `summary.sequence` in the local database. When node restart it resumes the stream by reading last processed checkpoint sequence number from the database and calls `CheckpointService::subscribe(last_processed_index)` to resume checkpoint stream.

The current implementation is not exactly perf optimized, for example each subscriber reads from database. This can be optimized later if more than a few consumers of this API are introduced.

Note that currently this API only works on validator, it will be available on full node once checkpoint sync is working.