[data model] enforce 250 KB size limit for Move objects
Don't allow the BCS bytes of a Move value object to be larger than 250 KB. This prevents performance issues from reading/writing/updating very large objects.

We have discussed a limit of 1-2MB before and may increase this limit accordingly, but want to start with something small to encourage folks to use the new dynamic fields feature (which is the recommended way for dealing with large collections or data).

We will enforce a separate, less stringent limit for package objects in a subsequent PR.