fix: monomorphize `verificationObligation`
The `VerificationObligation` is an accumulator for batch verification of signatures on the same message. It works predicated on the `batch_verify` method on the `VerifyingKey` trait, which as we know works in this way:
- performs a batch verification on these schemes that can (Ed25519, BLS),
- performs an iterated verification on these schemes that can't (Spec256k1)

The trait `ToObligationSignature`, despite comments to the contrary, was doing runtime detection of the call to `VerificaitonObligation::add_signature_and_public_key`, which took any signature as argument, converted some of them to an obligation (BLS), the others to an `Obligation::None` (the other two schemes, despite Ed25519 having a valid batch verificaiton semantic), and errored upon encountering any `Obligation::None`.
The reason for this isn't that we want to address verification generically, it is that we want to build accumulators (`VerificationObligation`) for the authority signature. If you happen to want to add a signature that's not of the authority's scheme (in our case, BLS), that's an error.

A simple way of detecting this statically is to make `VerificaitonObligation` monomorphic, making calling `add_signature_and_public_key` on anything but the authority's chosen scheme a compilation error. That's what this commit does.