Fix faulty test (#6476)
The problem was that the transfer_coin call transfers away a random
object without checking if it is a gas object. It was transfering away
object_ref_v1 instead of a random coin, which was making the later
make_transfer_object_transaction_with_wallet_context call fail.

(The right fix would be to improve the testing code to make it less
error prone but this is blocking CI right now).