[reconfig] Initiate reconfig every N checkpoints (#6894)
This PR adds the initial trigger for starting reconfiguration. It's done
at every N checkpoints.
An alternative is to initiate this at some fixed time. However we
currently don't have timestamp out of Narwhal yet, and it's
significantly more complicated to do this through time. For Wave 2 where
we are doing it every ~6 hours, time makes less sense because it's not
aligned on days anyway. We could just tune the frequency to match
roughly 6 hours when we ship to Wave 2.