[verify] Limit check for 0x0 addresses to publish command
Previously this check ran whenever a module was compiled, preventing
the compilation of modules with non-zero addresses, with the intention
of preventing the accidental re-publishing of an already published module.

However, this blocks uses of a compiled package apart from for
publishing such as verifying that it matches an on-chain package, or
running its tests.

Loosening the restriction by only making this check when publishing a
module (where it is relevant, and also double checked by the validator).

Test Plan:

```
sui$ cargo nextest run
sui$ cargo simtest
```