[gas] encapsulating object.storage_rebate to avoid accidental minting/burning
`Object.storage_rebate` is a particularly sensitive field because it corresponds directly to SUI that the user can claim. This was previously a public field, which made it a bit too easy to make mistakes when updating this field. Made the field private, and forced the API's for accessing it to provide the old version of the object, or insist that no object size change is occurring.