[checkpoints] Limit number of transactions per checkpoint
We had an issue in testnet where we created too large checkpoints and caused anemo to reject checkpoint messages.

While this particular incident was caused by a bug in checkpoint builder, it is also possible to naturally have a very large checkpoint, for example if narwhal was struggling for some time and created a large commit after that.

We need to limit number of transactions per checkpoint to avoid issues like that in the future.