[checkpoints] Correctly read last checkpoint on epoch boundary.
This PR changes how last checkpoint is read on epoch boundary. Previously we relied on internal builder table, `checkpoint_summary` to read last checkpoint. This is not correct however, since on the checkpoint boundary this might not be updated fully, also this table will be empty if new validator joins in the epoch.

In this PR we change it:
(1) `checkpoint_summary` is renamed and move to epoch store
(2) Across epoch boundary `certified_checkpoint` table is used to get last checkpoint of previous epoch.