[client/publish] Add ability to publish --with-unpublished-dependencies
If a package depends on other packages that haven't been published
yet, this flag offers a convenience to publish them as part of the
same transaction (or to include them in the base64 dump when
building).

This is not ideal from the perspective of code re-use, but does help
while wipes are more common.

This also has implications for source validation, which needs to
support validating packages that have other packages embedded in them:

- Now, on-chain modules in the root package could be matched against
  multiple source packages, so they are fetched all in one go.
- This also means there is no longer a 1:1 correspondence between
  numerical address and on-chain address, so the
  `LocalDependencyNotFound` error now identifies a package by its address.

Test Plan:

New source-validation and move integration tests:

```
$ cargo simtest
$ cargo nextest run
```