Pick (ccef7c9) [faucet] Write-ahead Log (#7305) (#7434)
Add a persistent store to the faucet that remembers all the requests it
has sent to full-node. Entries get removed when a definite response
comes back, and remain if not (i.e. if the faucet gives up on a coin
because of a timeout, and discards it).

This adds an extra layer of protection when full node goes down --
previously faucet would discard the coins to avoid equivocating them --
now those coins and the respective transactions will be remembered in
the WAL, and retried when the faucet is restarted, allowing us to
recover without discarding *or* equivocating coins.

Test Plan:

New unit tests for WAL:

```
$ cargo simtest
$ cargo nextest run
```

Add a sleep in the flow between when faucet adds a record to the WAL and
removes it, run a local validator:

```
sui$ cargo build
sui$ rm -rf ~/sui/
sui$ ~/sui/target/debug/sui genesis -f
sui$ ~/sui/target/debug/sui start
```

Faucet:

```
sui$ ~/sui/target/debug/sui-faucet \
     --max-request-per-second 10   \
     --write-ahead-log /tmp/faucet.wal
```

Send it a request:

```
$ set -l ADDRESS (yq -r .active_address ~/.sui/sui_config/client.yaml | sed 's/0x//' | string upper)
  curl --location --request POST 'http://127.0.0.1:5003/gas' \
       --header 'Content-Type: application/json' \
       --data-raw "{ \"FixedAmountRequest\": { \"recipient\": \"$ADDRESS\" } }"
```

Interrupt (^C) the faucet during the sleep, and restart it:

```
sui$ ~/sui/target/debug/sui-faucet \
     --max-request-per-second 10   \
     --write-ahead-log /tmp/faucet.wal
2023-01-10T22:29:48.297803Z  INFO sui_faucet::faucet::simple_faucet: Adding coin to gas pool coin_id=0x0898e1d2fd78550c7b546f5eb2c085d24563c685
2023-01-10T22:29:48.298162Z  INFO sui_faucet::faucet::simple_faucet: Adding coin to gas pool coin_id=0x3d5636a0ac75afbc482462bab9dc08e48fe9d70e
2023-01-10T22:29:48.298250Z  INFO sui_faucet::faucet::simple_faucet: Retrying txn from WAL. uuid=b314c6a4-92fb-477a-badf-228abaa41012 recipient=0xfb560f0ebd8a153fc1a8d2fa647303ff2281512d coin_id=0x479db03f2391ef59f6803fd0605102ed9fbe935f
2023-01-10T22:29:48.298289Z  INFO sui_faucet::faucet::simple_faucet: Adding coin to gas pool coin_id=0x5bb3e35fdf29d2e27913217a332f1ad5ef721b13
... [snip] ...
2023-01-10T22:29:48.298432Z  INFO sui_faucet::faucet::simple_faucet: Adding coin to gas pool coin_id=0xf051714b916d9bab26c44f446792efc9d6b4dadc
2023-01-10T22:29:48.299236Z  INFO sui_faucet::faucet::simple_faucet: PaySui transaction in faucet. tx_digest=EV55z1w7dK6CFbWA4b9ABqVQjCyQF8Lim445yK9RuQeo recipient=0xfb560f0ebd8a153fc1a8d2fa647303ff2281512d coin_id=0x479db03f2391ef59f6803fd0605102ed9fbe935f uuid=b314c6a4-92fb-477a-badf-228abaa41012
2023-01-10T22:29:48.317487Z  INFO sui_faucet::faucet::simple_faucet: About to slow things down coin_id=0x479db03f2391ef59f6803fd0605102ed9fbe935f
2023-01-10T22:29:58.320584Z  INFO sui_faucet::faucet::simple_faucet: Got producer lock and recycling coin uuid=b314c6a4-92fb-477a-badf-228abaa41012 coin_id=0x479db03f2391ef59f6803fd0605102ed9fbe935f
2023-01-10T22:29:58.320638Z  INFO sui_faucet::faucet::simple_faucet: Recycled coin uuid=b314c6a4-92fb-477a-badf-228abaa41012 coin_id=0x479db03f2391ef59f6803fd0605102ed9fbe935f
2023-01-10T22:29:58.320919Z  INFO sui_faucet: listening on 127.0.0.1:5003

```

It will retry the transactions that are still in the WAL.