Reject transaction if any input object is overloaded (#7438)
It is possible for users to create many transactions with the same
shared input object, in a rate that is faster than the rate Sui can
execute them serially. This PR adds a basic rate limiting mechanism:
reject signing or submitting new transactions, if one of its input
object ID has > 1000 pending transactions. This scheme is not
bulletproof, but should help the case when users are unintentionally
generating too much load.