[reconfig] Add advance epoch safe mode (#7555)
This PR adds a simplified version of `advance_epoch` that only updates
the epoch ID.
In the execution_engine, if we failed to execute the normal
advance_epoch, we fall back to execute the simple one.
And if this ever happened, we memorized it through a bool safe_mode and
will forever be in safe_mode.
Also added logging and metric so that we could monitor it.