[object] Store package version in object
Soon it will no longer be the case that a package's version is always
`1`, because:

- Framework Upgrades will write new versions of the framework at
  increasing versions in the store.
- Package Upgrades will pun the package's version in storage with its
  version in the chain of upgrades (Even though subsequent versions
  will exist at distinct ObjectIDs).

This PR retires `PACKAGE_VERSION` and stores the package version in
`MovePackage` instead (although initially, it will still always be 1).

This change also means that it is no longer possible to ascertain a
package's version, when it is passed in as an input object.

Breaking Changes:

- Call transactions that reference a non-existent package will produce
  a "Missing Input Object" error.  Because the query for the package
  is not fixed to a particular version, `VersionNumber::MAX` is
  returned as a sentinel value for the version it is missing at.

Test Plan:

```
$ cargo nextest run
```

Stack:

- #7597