[move/call] Accept Package by ID, not Ref
Update the `MoveCall` transaction type to accept its `package` by
`ObjectID` and not `ObjectRef` (i.e. drop the version and digest).

This step is required to support framework upgrades.  A signed
transaction that calls an entry function in the framework package
should continue to be valid after a framework upgrade.  This is not
true if the transaction has to specify the version of the package
being called, because this gets incremented during the upgrade.

There is not a lot of substance to this change, because we were
already completely ignoring a package's version and digest on the
validator -- the version is not used for reads because those are
handled by the Module Resolver whose keys are `(AccountAddress,
ModuleName)` (i.e. no version), and the digest is not checked by the
transaction input checker.

Breaking Changes:

RPC endpoints for move calls (both Open RPC and JSON RPC), now accept
a package by ID, not Ref (`(ID, Version, Digest)`).  This means SDKs
no longer have to resolve a package to its ref on the client, to build
a `MoveCall` transaction.

Test Plan:

```
$ cargo nextest run
$ pnpm sdk install
$ pnpm sdk build
$ pnpm sdk test:unit
$ pnpm sdk prepare:e2e
$ pnpm sdk test:e2e
```